fixing and refactoring tests
