+++ tests/data/source/test_src_data.csv
@@ -1,0 +1,10 @@
+Data Year,ID,Property Name,Reporting Status,Address,ZIP Code,Chicago Energy Rating,Exempt From Chicago Energy Rating,Community Area,Primary Property Type,Gross Floor Area - Buildings (sq ft),Year Built,# of Buildings,Water Use (kGal),ENERGY STAR Score,Electricity Use (kBtu),Natural Gas Use (kBtu),District Steam Use (kBtu),District Chilled Water Use (kBtu),All Other Fuel Use (kBtu),Site EUI (kBtu/sq ft),Source EUI (kBtu/sq ft),Weather Normalized Site EUI (kBtu/sq ft),Weather Normalized Source EUI (kBtu/sq ft),Total GHG Emissions (Metric Tons CO2e),GHG Intensity (kg CO2e/sq ft),Latitude,Longitude,Location,Row_ID,Wards,Community Areas,Zip Codes,Census Tracts,Historical Wards 2003-2015
+2020,256458,United Center Office Building,Not Submitted,1901 W Madison St,60612,0.0,False,,,,,,,,,,,,,,,,,,,41.88125398,-87.67448493,"(41.88125398, -87.67448493)",2020-256458,46.0,29.0,21184.0,90.0,41.0
+2020,100856,United Center,Not Submitted,1901 W Madison St,60612,0.0,False,NEAR WEST SIDE,,2289000.0,,,,,,,,,,,,,,,,41.88067672,-87.67418207,"(41.88067672, -87.67418207)",2020-100856,46.0,29.0,21184.0,90.0,41.0
+2017,100856,United Center,Submitted,1901 West Madison St.,60612,,,NEAR WEST SIDE,Indoor Arena,960000.0,1994.0,2.0,,,111594203.0,8070756.0,,,,124.7,373.8,125.2,374.4,21061.9,21.9,41.88067672,-87.67418207,"(41.88067672, -87.67418207)",2017-100856,46.0,29.0,21184.0,90.0,41.0
+2015,100856,United Center,Submitted,1901 West Madison St.,60612,,,NEAR WEST SIDE,Indoor Arena,960000.0,1994.0,1.0,,,11479770.0,12102665.0,,,,24.6,50.8,,,2949.0,3.1,41.88067672,-87.67418207,"(41.88067672, -87.67418207)",2015-100856,46.0,29.0,21184.0,90.0,41.0
+2018,100856,United Center,Submitted,1901 West Madison St.,60612,4.0,False,NEAR WEST SIDE,Indoor Arena,960000.0,1994.0,2.0,12929.0,,11345016.0,9342860.4,,,,21.5,43.3,21.6,43.4,2383.6,2.5,41.88067672,-87.67418207,"(41.88067672, -87.67418207)",2018-100856,46.0,29.0,21184.0,90.0,41.0
+2016,100856,United Center,Submitted,1901 West Madison St.,60612,,,NEAR WEST SIDE,Indoor Arena,960000.0,1994.0,1.0,,,113634101.0,8970953.0,,,,127.7,381.5,128.5,382.3,21421.7,22.3,41.88067672,-87.67418207,"(41.88067672, -87.67418207)",2016-100856,46.0,29.0,21184.0,90.0,41.0
+2019,100856,United Center,Submitted,1901 W Madison St,60612,2.0,False,NEAR WEST SIDE,Indoor Arena,960000.0,1994.0,2.0,206239.0,,102653875.6,15169580.2,,,,122.7,316.0,122.4,,17883.7,18.6,41.88067672,-87.67418207,"(41.88067672, -87.67418207)",2019-100856,46.0,29.0,21184.0,90.0,41.0
+2021,100856,United Center,Not Submitted,1901 W Madison St,60612,0.0,False,NEAR WEST SIDE,Indoor Arena,2289000.0,1994.0,1.0,,,,,,,,,,,,,,41.88067672,-87.67418207,"(41.88067672, -87.67418207)",2021-100856,46.0,29.0,21184.0,90.0,41.0
+2021,256458,United Center Office Building,Not Submitted,1901 W Madison St,60612,0.0,False,,,,,,,,,,,,,,,,,,,41.88125398,-87.67448493,"(41.88125398, -87.67448493)",2021-256458,46.0,29.0,21184.0,90.0,41.0
